+++ R/split_var.R
@@ -18,8 +18,10 @@
 #' @inheritParams group_var
 #' @inheritParams rec
 #'
-#' @return A grouped variable with equal sized groups. If \code{x} is a data
-#'         frame, only the grouped variables will be returned.
+#' @return A grouped variable with equal sized groups. If \code{x} is a data frame,
+#'         for \code{append = TRUE}, \code{x} including the grouped variables
+#'         as new columns is returned; if \code{append = FALSE}, only
+#'         the grouped variables will be returned.
 #'
 #' @details \code{split_var()} splits a variable into equal sized groups, where the
 #'            amount of groups depends on the \code{groupcount}-argument. Thus,
@@ -60,12 +62,15 @@
 #'
 #' @importFrom stats quantile
 #' @export
-split_var <- function(x, ..., groupcount, as.num = FALSE, val.labels = NULL, var.label = NULL, inclusive = FALSE, suffix = "_g") {
+split_var <- function(x, ..., groupcount, as.num = FALSE, val.labels = NULL, var.label = NULL, inclusive = FALSE, append = FALSE, suffix = "_g") {
   # evaluate arguments, generate data
   .dots <- match.call(expand.dots = FALSE)$`...`
   .dat <- get_dot_data(x, .dots)
 
   if (is.data.frame(x)) {
+
+    # remember original data, if user wants to bind columns
+    orix <- tibble::as_tibble(x)
 
     # iterate variables of data frame
     for (i in colnames(.dat)) {
@@ -86,6 +91,9 @@ split_var <- function(x, ..., groupcount, as.num = FALSE, val.labels = NULL, var
     if (!is.null(suffix) && !sjmisc::is_empty(suffix)) {
       colnames(x) <- sprintf("%s%s", colnames(x), suffix)
     }
+
+    # combine data
+    if (append) x <- dplyr::bind_cols(orix, x)
   } else {
     x <- split_var_helper(
       x = .dat,